components/risk-engine/pyproject.toml
# TODO: to be updated
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = ""
authors = ["b-labs"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
sqlalchemy = "^2.0.20"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
httpx = "^0.25.0"
psycopg2-binary = "^2.9.7"
alembic = "^1.12.0"
datadog-api-client = "^2.18.0"
pre-commit = "^3.4.0"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.6"
mixer = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
